refactor(runtime): SceneSerializer.cpp becomes five files
1,804 lines covering entity references, instance staging, level save/load,
blueprint preparation and disk I/O.

  SceneSerializerRefs.cpp        entity references, moving entities between scenes
  SceneSerializerInstances.cpp   staging and committing one instance
  SceneSerializerLevel.cpp       saving and loading a whole level
  SceneSerializerBlueprint.cpp   preparing a blueprint, placing, re-expanding
  SceneSerializerFiles.cpp       reading and writing levels on disk

Two things genuinely crossed every seam and are now internal headers rather than
one anonymous-namespace block. SceneSerializerContext.hpp holds the thread-local
serialization context — the reference hooks, Save, Load and every instance path
read and write it, so it is `inline thread_local` rather than a copy per TU.
SceneSerializerInstances.hpp holds StagedInstance, AdoptionSet and the staging
functions, because Load, PlaceInstance and ReexpandInstance all stage before
they commit.

StageInstance's default argument moves to its declaration, where it belongs now
that there is one. SanitizeNonFinite has a single caller in Save and stays
anonymous there; StripSerializable and LeafName likewise, in the instances file.

Build clean, ctest 14/14.
