+++ modules/Runtime/src/SceneSerializerRefs.cpp
@@ -1,0 +1,321 @@
+/* Copyright (c) 2025 Francisco Vivas Puerto (aka "DaFrancc"). */
+#include <Assisi/Runtime/SceneSerializer.hpp>
+
+#include <Assisi/Core/Assert.hpp>
+#include <Assisi/Core/AssetSystem.hpp>
+#include <Assisi/Core/BitStream.hpp>
+#include <Assisi/Core/Logger.hpp>
+#include <Assisi/Core/Reflect/BinaryCodec.hpp>
+#include <Assisi/Core/Reflect/ComponentRegistry.hpp>
+#include <Assisi/ECS/BlueprintMember.hpp>
+#include <Assisi/Runtime/Blueprint.hpp>
+#include <Assisi/Runtime/EditorOnly.hpp>
+#include <Assisi/Runtime/Hierarchy.hpp>
+#include <Assisi/Runtime/NameComponent.hpp>
+#include <Assisi/Runtime/Naming.hpp>
+
+#include <cmath>
+#include <cstdint>
+#include <format>
+#include <fstream>
+#include <map>
+#include <memory>
+#include <optional>
+#include <span>
+#include <stdexcept>
+#include <string>
+#include <unordered_map>
+#include <unordered_set>
+#include <vector>
+
+#include "SceneSerializerContext.hpp"
+
+// ---------------------------------------------------------------------------
+// Entity references, and moving entities between scenes.
+// ---------------------------------------------------------------------------
+
+namespace Assisi::Runtime
+{
+
+std::string_view Describe(LevelError error)
+{
+    switch (error)
+    {
+    case LevelError::FileUnreadable:
+        return "the file could not be read";
+    case LevelError::MalformedJson:
+        return "the file is not readable JSON";
+    case LevelError::UnsupportedVersion:
+        return "the file is a version this build does not read";
+    case LevelError::NoInstanceTable:
+        return "the file places blueprint instances and this load has nowhere to put them";
+    case LevelError::MissingName:
+        return "an entity or instance in the file has no name";
+    case LevelError::MissingSource:
+        return "an instance in the file names no source";
+    case LevelError::InvalidName:
+        return "the file gives something a name that is not usable as one";
+    case LevelError::DuplicateName:
+        return "two things in the file answer to one name";
+    case LevelError::NonUniformScale:
+        return "an instance has a non-uniform scale";
+    case LevelError::BlueprintUnusable:
+        return "an instance names a blueprint that will not load";
+    case LevelError::UnresolvedReference:
+        return "a reference names something the file does not declare";
+    case LevelError::MalformedComponent:
+        return "a component field holds something the engine cannot read";
+    case LevelError::ContextBusy:
+        return "a serialization context is already active on this thread";
+    case LevelError::InstanceNotLive:
+        return "no such instance is live";
+    case LevelError::NameAlreadyLive:
+        return "an instance of that name is already live in this world";
+    }
+    return "the file cannot be used";
+}
+
+// ---------------------------------------------------------------------------
+// Public context accessors (called from component serialize/addToScene lambdas)
+// ---------------------------------------------------------------------------
+
+nlohmann::json SceneSerializer::EntityToRef(ECS::Entity entity)
+{
+    // Null first, and unconditionally: every mode spells "no target" the same way,
+    // and none of the lookups below has a meaningful answer for it.
+    if (entity == ECS::NullEntity)
+        return nullptr;
+
+    // Raw-entity context wins: identity mapping, keyed by slot AND generation.
+    //
+    // Round-6 M11: this used to return the bare slot index, dropping the
+    // generation. A ref captured to an entity that was later destroyed then
+    // resolved to whatever new entity reused the slot — silently, because the
+    // reused slot is perfectly alive, so no liveness check can catch it. Packing
+    // the generation in lets RefToEntity below reject the stale ref instead.
+    if (s_rawContextScene != nullptr)
+        return EntityKey(entity.index, entity.generation);
+
+    if (!s_context)
+        return nullptr;
+
+    const uint64_t key = EntityKey(entity.index, entity.generation);
+
+    if (s_context->mode == SerializationContext::RefMode::SetIndices)
+    {
+        const auto it = s_context->entityToIndex.find(key);
+        // Out of the moved set: ~0ull, which RefToEntity reads as out of range and
+        // nulls. TransferEntities warns about each of these before it gets here.
+        return it != s_context->entityToIndex.end() ? nlohmann::json(it->second) : nlohmann::json(~0ull);
+    }
+
+    const auto it = s_context->entityToName.find(key);
+    return it != s_context->entityToName.end() ? nlohmann::json(it->second) : nlohmann::json(nullptr);
+}
+
+ECS::Entity SceneSerializer::RefToEntity(const nlohmann::json &value)
+{
+    if (value.is_null())
+        return ECS::NullEntity;
+
+    // Raw-entity context wins: the value is a packed (slot, generation) key. The
+    // paired restore revives entities at their original handle, so in the intended
+    // flow the generation matches exactly. When it does not, the slot has been
+    // recycled by an unrelated entity and the ref is stale — resolve to null
+    // rather than silently redirecting onto whoever moved in (round-6 M11).
+    if (s_rawContextScene != nullptr)
+    {
+        if (!value.is_number_unsigned())
+            return ECS::NullEntity;
+        const uint64_t    key     = value.get<uint64_t>();
+        const auto        slot    = static_cast<uint32_t>(key & 0xFFFFFFFFull);
+        const auto        wantGen = static_cast<uint32_t>(key >> 32);
+        const ECS::Entity live    = s_rawContextScene->EntityAt(slot);
+        return live.generation == wantGen ? live : ECS::NullEntity;
+    }
+
+    if (!s_context)
+        return ECS::NullEntity;
+
+    if (s_context->mode == SerializationContext::RefMode::SetIndices)
+    {
+        if (!value.is_number_unsigned())
+            return ECS::NullEntity;
+        const uint64_t index = value.get<uint64_t>();
+        return index < s_context->indexToEntity.size() ? s_context->indexToEntity[static_cast<std::size_t>(index)]
+                                                       : ECS::NullEntity;
+    }
+
+    // A file ref is a name. Anything else is a v1 file or a hand-edit that meant a
+    // position — both of which would resolve to *some* entity if we guessed, which
+    // is the failure names exist to remove. Record it and let Load refuse.
+    if (!value.is_string())
+    {
+        s_context->unresolvedRefNames.push_back(value.dump());
+        return ECS::NullEntity;
+    }
+
+    // A leading slash means "the file that wrote this", which for a reference in
+    // the outermost file is that file — so it addresses the same names a plain
+    // one does. Inside an instance the prefixing already resolved the difference
+    // (Runtime::Blueprint's QualifyName), and this is what makes the two agree.
+    auto name = value.get<std::string>();
+    if (!name.empty() && name.front() == '/')
+        name.erase(0, 1);
+
+    const auto it = s_context->nameToEntity.find(name);
+    if (it == s_context->nameToEntity.end())
+    {
+        s_context->unresolvedRefNames.push_back(std::move(name));
+        return ECS::NullEntity;
+    }
+
+    // Claimed but mapped at nothing: a member an instance removed. That is a
+    // legitimate thing for a file to say, so the reference nulls with a warning
+    // rather than refusing the file the way an unknown name does.
+    if (it->second == ECS::NullEntity)
+        Core::Log::Warn("SceneSerializer: a reference names '{}', which its instance removed — left null.", name);
+
+    return it->second;
+}
+
+// ---------------------------------------------------------------------------
+// TransferEntities (entity migration)
+// ---------------------------------------------------------------------------
+
+std::vector<ECS::Entity> SceneSerializer::TransferEntities(ECS::Scene &src, ECS::Scene &dst,
+                                                           std::span<const ECS::Entity> entities)
+{
+    if (s_context || s_rawContextScene != nullptr)
+    {
+        Core::Log::Error("TransferEntities: a serialization context is already active on this thread.");
+        return {};
+    }
+    if (entities.empty())
+        return {};
+
+    ScopedContextReset guard;
+    s_context       = SerializationContext{};
+    s_context->mode = SerializationContext::RefMode::SetIndices;
+
+    // Source half of the remap: each migrated entity → its index within the set.
+    // A ref to any entity NOT in this map returns nullopt from EntityToIndex,
+    // serializes as ~0ull, and resolves to NullEntity on the destination side —
+    // exactly the "null a ref that leaves the set" behaviour we want.
+    for (uint32_t i = 0; i < entities.size(); ++i)
+    {
+        const ECS::Entity e = entities[i];
+        s_context->entityToIndex[EntityKey(e.index, e.generation)] = i;
+    }
+
+    const auto &registry = Core::Reflect::ComponentRegistry::Instance();
+
+    // Diagnose refs that will be dropped, before serialize silently loses them.
+    // Walk each migrated entity's reflected EntityRef fields; a non-null handle
+    // that is not itself in the migrated set is about to become null.
+    for (const ECS::Entity e : entities)
+    {
+        for (const Core::Reflect::ComponentMeta *meta : registry.SerializableComponents())
+        {
+            const void *comp = meta->getByEntity(&src, e.index, e.generation);
+            if (comp == nullptr)
+                continue;
+            for (const Core::Reflect::FieldMeta &field : meta->fields)
+            {
+                if (field.type != Core::Reflect::FieldType::EntityRef || field.transient)
+                    continue;
+                const auto ref = *reinterpret_cast<const ECS::Entity *>(static_cast<const char *>(comp) +
+                                                                        field.offset);
+                if (ref == ECS::NullEntity)
+                    continue;
+                if (!s_context->entityToIndex.contains(EntityKey(ref.index, ref.generation)))
+                {
+                    Core::Log::Warn("Migrate: {}::{} on entity (index {}, gen {}) references entity "
+                                    "(index {}, gen {}) outside the migrated set — it will be null in "
+                                    "the destination.",
+                                    meta->name, field.name, e.index, e.generation, ref.index, ref.generation);
+                }
+            }
+        }
+    }
+
+    // Pass 1: serialize every migrated component while the source map is live, so
+    // in-set EntityRefs capture their set index. Held per entity for pass 3.
+    struct CapturedComponent
+    {
+        const Core::Reflect::ComponentMeta *meta;
+        nlohmann::json                      data;
+    };
+    std::vector<std::vector<CapturedComponent>> captured(entities.size());
+    for (std::size_t i = 0; i < entities.size(); ++i)
+    {
+        const ECS::Entity e = entities[i];
+        for (const Core::Reflect::ComponentMeta *meta : registry.SerializableComponents())
+        {
+            if (const void *comp = meta->getByEntity(&src, e.index, e.generation))
+                captured[i].push_back({meta, meta->serialize(comp)});
+        }
+    }
+
+    // Pass 2: create the destination entities and record the destination half of
+    // the remap. All created first, so a ref to a sibling that appears later in
+    // the set still resolves in pass 3 (same forward-ref handling as Load).
+    std::vector<ECS::Entity> created;
+    created.reserve(entities.size());
+    s_context->indexToEntity.reserve(entities.size());
+    for (std::size_t i = 0; i < entities.size(); ++i)
+    {
+        const ECS::Entity d = dst.Create();
+        created.push_back(d);
+        s_context->indexToEntity.push_back(d);
+    }
+
+    // Pass 3: deserialize each captured component into its destination entity.
+    // IndexToEntity now maps in-set refs onto the created handles.
+    for (std::size_t i = 0; i < entities.size(); ++i)
+    {
+        for (const CapturedComponent &c : captured[i])
+        {
+            // The data came out of Save on the *source* scene a moment ago, so a
+            // refusal here means the codec disagrees with itself rather than that
+            // a file is wrong. Loud, and the migration continues: the entity has
+            // already moved and there is nothing to roll back to.
+            if (c.meta->addToScene &&
+                !c.meta->addToScene(&dst, created[i].index, created[i].generation, c.data))
+            {
+                Core::Log::Error("SceneSerializer: migrating '{}' lost a component the source scene had "
+                                 "written — this is an engine bug, not a bad file.",
+                                 c.meta->name);
+            }
+        }
+    }
+
+    // The originals leave the source scene. Structural, so defer to the source's
+    // own FlushDestroyed (end of its frame) rather than mutating mid-iteration —
+    // the caller tears down their Jolt bodies, which the ECS destroy does not.
+    for (const ECS::Entity e : entities)
+        src.Destroy(e);
+
+    return created;
+}
+
+// ---------------------------------------------------------------------------
+// ScopedRawEntityContext
+// ---------------------------------------------------------------------------
+
+SceneSerializer::ScopedRawEntityContext::ScopedRawEntityContext(ECS::Scene &scene)
+{
+    // One context per thread. Nesting raw-in-raw or raw-in-Save/Load is a bug:
+    // the mappings are incompatible and the flat pointer/optional can't stack.
+    ASSISI_ASSERT(s_rawContextScene == nullptr && !s_context,
+                  "SceneSerializer: a serialization context is already active on this thread "
+                  "(raw-entity and Save/Load contexts are mutually exclusive and non-reentrant).");
+    s_rawContextScene = &scene;
+}
+
+SceneSerializer::ScopedRawEntityContext::~ScopedRawEntityContext()
+{
+    s_rawContextScene = nullptr;
+}
+
+} // namespace Assisi::Runtime
